+++ CHANGELOG.md
@@ -5,7 +5,19 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
-## [0.2.0] - 2019-09-24
+## [v0.3.0] - 2019-12-01
+### Added
+- Add Events.OnEnterVehicle callback. Saves coordinates of the player's entry into the vehicle to {{dd-mm-yy_h-i}}_cmd.txt log file.
+- Add Events.OnExitVehicle callback. Saves player's exit coordinates from vehicle to {{dd-mm-yy_h-i}}_cmd.txt log file.
+- Add Events.EveryHours callback. Makes saving the character to {{dd-mm-yy_h-i}}_player.txt log file every one ingame hour.
+
+### Changed
+- Events can be turned on or off in the configuration of the LogExtender object.
+
+### Fixed
+- Removed levelup entries from the character’s creation window until the server is fully connected.
+
+## [v0.2.0] - 2019-09-24
 ### Added
 - Add level to dump player stats.
 - Start writing changelog.
@@ -28,7 +40,8 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Add basic implementation.
 
-[Unreleased]: https://github.com/gorcon/rcon-cli/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/gorcon/rcon-cli/compare/v0.3.0...HEAD
+[v0.3.0]: https://github.com/gorcon/rcon-cli/compare/v0.2.0...v0.3.0
 [v0.2.0]: https://github.com/gorcon/rcon-cli/compare/v0.1.2...v0.2.0
 [v0.1.2]: https://github.com/gorcon/rcon-cli/compare/v0.1.1...v0.1.2
 [v0.1.1]: https://github.com/gorcon/rcon-cli/compare/v0.1.0...v0.1.1